stock/requirements.txt
crewai
pydantic
python-dotenv
crewai-tools
